Enable language switcher for both contact views
